Add new variants of set_user and reset_user which allow a token to be set initially and if so required to be present and match at reset time.
Document this change as well as some general improvements to
documentation regarding the block_superuser custom setting.

Finally, if block_log_statement has been turned off, don't
change log_statement at all. It is silly to change it since
the user could easily turn it off or on, and it is probably
annoying to have it turned on by default in those cases.