Add post-set_user hooks
This commit introduces two hooks which allow another extension to do something
(and return) immediately after `set_user` and `reset_user` are called.

There is a new make target, `make install-headers`, which is responsible for
exposing the hook to other extensions. The target is prepended to the
`install` target by default.

To use the post-set_user hooks in an extension:
1) Add '-I$(includedir)' to CPPFLAGS
2) #include set_user.h in whichever file implements the hook
3) Register the 'post_reset_user_hook' and 'post_set_user_hook' with a local
implementation