Fix const warning and log formatting
GetConfigOption returns a const char * and it was being handled without casting,
since removal of pstrdup in the last version of the patch. Fix this by declaring
the variable const char * per other examples in postgres.

Additionally account for typical use of whitespace at the end of
`log_line_prefix` variables by not adding an extra whitespace character between
the old value and the audit tag.