Use rendezvous variables for set_user post hooks
This patch changes the set_user hooks so that they're held in the rendezvous
variable hash map as function pointers, rather than in the set_user .so. This
way, we don't have to ensure that set_user comes before any extension that
implements them -- load order doesn't matter. Just query the hash map during the
hook execution and if there are hooks implemented, call them.