Update authorized uid assertion logic
pgaudit/set_user@b74c1d8 introduced a preprocessor flag to disable
asserts on AuthorizedUserId initialization, but missed the case where
assertions are off altogether and no convencience flag is set.

This commit fixes that issue by simplifying the logic a bit and cleaning
up some of the associated code.