Properly process next object_access_hook
Commit f56384c5767d ("Introduce object_access hook to set_user") adds a
new object_access_hook to the queue.

The hook is only called after a user initiates `set_user[u]()`; however,
it also skips other hooks if not.

To fix this, move handling of next hooks out of the conditional, so we
don't block in the even that `set_user[u]()` has not been called.