Update CI/local test infrastructure. (#93)
The CI tests were awkward to run locally and were not documented for local use. In addition, we were at the mercy of the container providers to release beta versions. Two weeks after 18beta1 was released there was still no container.

Instead reuse the test framework from pgAudit modified for the needs of set_user. This framework is easy to run locally or in CI and automatically pulls the most recent build of each major version without needing to change the workflow.

As a bonus the tests run about twice as fast. 1m vs 2m is not game changing but still nice.