Remove out-of-date header version and hard to maintain version comments.
The version in the header comment was out of date and there was no clear procedure for keeping it up to date. Instead just strip the version number so it does not need to be updated for each version.

Also remove some comments that said nothing was added for a particular version. Many versions are missing from these comments and it is not too clear when they should be added. It seems better to just note versions where something was added.