Remove dead branch. (#94)
The compiler warns that this branch is unreachable:

src/set_user.c:241:26: warning: the comparison will always evaluate as 'true' for the address of 'data' will never be NULL [-Waddress]
  241 |                 else if (!NameStr(procStruct->proname))
      |                          ^
In file included from /usr/include/postgresql/17/server/postgres.h:45,
                 from src/set_user.c:28:
/usr/include/postgresql/17/server/c.h:742:25: note: 'data' declared here
  742 |         char            data[NAMEDATALEN];
      |                         ^~~~

This warning goes all the way back to PG12 so the code has not been valid for at that long. One solution would be to change the branch to:

else if (NameStr(procStruct->proname)[0] == '\0')

However, it is not clear that this will ever be true. If SearchSysCache1() returned successfully presumably the name got filled in.

Overall it seems best to just remove the branch since it leaves us no worse off than we were before. This will allow us to enable error on warnings since this is the only warning currently.

Reviewed by David Christensen (@pgguru).